@@ -14,25 +14,27 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "2.4.1",
-    "@docusaurus/plugin-google-gtag": "^2.4.1",
-    "@docusaurus/preset-classic": "2.4.1",
-    "@mdx-js/react": "^1.6.22",
+    "@docusaurus/core": "3.4.0",
+    "@docusaurus/plugin-google-gtag": "^3.4.0",
+    "@docusaurus/preset-classic": "3.4.0",
+    "@mdx-js/react": "^3.0.0",
+    "clsx": "^2.0.0",
+    "prism-react-renderer": "^2.3.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "@mdx-js/react": "^3.0.0",
     "autoprefixer": "^10.4.7",
-    "clsx": "^1.1.1",
     "core-js": "^3.27.2",
     "mobx": "^6.8.0",
     "postcss": "^8.4.14",
-    "prism-react-renderer": "^1.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
     "react-github-btn": "^1.3.0",
     "react-icons": "^4.3.1",
     "styled-components": "^5.3.6",
-    "tailwindcss": "^3.0.24"
+    "tailwindcss": "^3.4.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "2.4.1",
+    "@docusaurus/module-type-aliases": "3.4.0",
+    "@docusaurus/types": "3.4.0",
     "eslint": "^8.45.0",
     "eslint-plugin-react": "^7.32.2"
   },
@@ -47,5 +49,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "engines": {
+    "node": ">=18.0"
   }
 }
